+++ packages/veritone-widgets/CHANGELOG.md
@@ -117,3 +117,18 @@ const MyPage = () => (
 
 ## 5.1.0
 * Added Engine Selection widget
+
+## 5.2.0
+* Added MediaPlayer component.
+* Fixed Redux devtools extension crashing.
+* Engine Selection widget
+  *  Fix initial selected engines not reinitializing across widgets.
+  *  Optimized search by sharing search results state across widgets.
+
+## 5.2.1
+* Moved Video-React from a peer dependency to a dependency
+* Bumped Redux to 4.0
+
+## 5.2.2
+* Bumped internal veritone-redux-common dependency.
+* Import `fetchGraphQLApi` from veritone-redux-common and remove own internal version.